# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyDocs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/docs/types": {
            "get": {
                "tags": [
                    "Tipos de Documentos"
                ],
                "summary": "Lista todos os tipos de documentos",
                "description": "Recupera todos os tipos de documentos existentes",
                "operationId": "types_index",
                "responses": {
                    "200": {
                        "description": "Tipos de documentos recuperados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os tipos de documentos existentes foram recuperados com sucesso."
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "types": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "documento": {
                                                        "type": "string",
                                                        "example": "Título do Tipo"
                                                    },
                                                    "sigla": {
                                                        "type": "string",
                                                        "example": "Sigla do Tipo"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tipos de Documentos"
                ],
                "summary": "Cadastra um novo tipo de documento",
                "description": "Cria um novo tipo de documento com base nos dados fornecidos",
                "operationId": "types_store",
                "requestBody": {
                    "description": "Informações do tipo de documento",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "abbreviation"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Título do Tipo"
                                    },
                                    "abbreviation": {
                                        "type": "string",
                                        "example": "Sigla do Tipo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de documento cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O tipo de documento foi cadastrado com sucesso."
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Título do Tipo"
                                                },
                                                "abbreviation": {
                                                    "type": "string",
                                                    "example": "Sigla do Tipo"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "tipo": {
                                            "properties": {
                                                "documento": {
                                                    "type": "string",
                                                    "example": "Título do Tipo"
                                                },
                                                "sigla": {
                                                    "type": "string",
                                                    "example": "Sigla do Tipo"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Problema ao cadastrar o tipo de documento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Houve um problema ao cadastrar o tipo de documento :("
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}